FEM - Functions

Index

PeriLab.Solver_Manager.FEM.create_module_specifics

PeriLab.Solver_Manager.FEM.create_module_specifics

PeriLab.Solver_Manager.FEM.find_jl_files

PeriLab.Solver_Manager.FEM.find_module_files

PeriLab.Solver_Manager.FEM.force_densities

PeriLab.Solver_Manager.FEM.include_files

FEM

PeriLab.Solver_Manager.FEM.create_module_specifics — Method

create_module_specifics(name::String, module_list::Dict{String,AbstractString}(),specifics::Dict{String,String}(), values::Tuple)

Searches for a specific function within a list of modules and calls that function if found.

This function iterates over a list of modules specified in module_list and looks for a module-specific function specified in the specifics dictionary. If the module and function are found, it calls that function with the provided values tuple.

Arguments

name::String: The name to match against the module names.

module_list::Dict{String, AbstractString}: A dictionary of module names mapped to abstract strings.

specifics::Dict{String, String}: A dictionary specifying the module-specific function to call for each module.

values::Tuple: A tuple of values to be passed as arguments to the module-specific function.

Example

```julia modulelist = Dict("Module1" => "Module1Name", "Module2" => "Module2Name") specifics = Dict("Module1Name" => "module1function", "Module2Name" => "module2function") values = (arg1, arg2) createmodulespecifics("Module1Name", modulelist, specifics, values)

PeriLab.Solver_Manager.FEM.create_module_specifics — Method

create_module_specifics(name::String, module_list::Dict{String,AbstractString}(),specifics::Dict{String,String}())
# Returns: the function itself

PeriLab.Solver_Manager.FEM.find_jl_files — Method

find_jl_files(directory::AbstractString)

Recursively find Julia files (.jl) in a directory.

This function recursively searches for Julia source files with the ".jl" extension in the specified directory and its subdirectories. It returns a vector of file paths for all the found .jl files.

Arguments

directory::AbstractString: The directory in which to search for .jl files.

Returns

A vector of strings, where each string is a file path to a .jl file found in the specified directory and its subdirectories.

Example

```julia jlfiles = findjlfiles("/path/to/modules") for jlfile in jlfiles println("Found Julia file: ", jlfile) end

PeriLab.Solver_Manager.FEM.find_module_files — Method

find_module_files(directory::AbstractString, specific::String)

Search for Julia modules containing a specific function in a given directory.

This function searches for Julia modules (files with .jl extension) in the specified directory and checks if they contain a specific function. It returns a list of dictionaries where each dictionary contains the file path and the name of the module where the specific function is found.

Arguments

directory::AbstractString: The directory to search for Julia modules.

specific::String: The name of the specific function to search for.

Returns

An array of dictionaries, where each dictionary has the following keys:

"File": The file path to the module where the specific function is found.

"Module Name": The name of the module where the specific function is found.

Example

```julia result = findmodulefiles("/path/to/modules", "myfunction") for moduleinfo in result println("Function found in module: ", moduleinfo["Module Name"]) println("Module file path: ", moduleinfo["File"]) end

PeriLab.Solver_Manager.FEM.force_densities — Method

force_densities(datamanager, nodes)

Computes the force densities from the FEM nodes.

Arguments

datamanager::Module: The datamanager.

nodes::Vector{Int64}: FEM nodes.

Returns

PeriLab.Solver_Manager.FEM.include_files — Method

include_files(module_list::Vector{Any})

Include files specified in a list of modules.

This function iterates over a list of modules and includes the files specified in each module's "File" key.

Arguments

module_list::Vector{Any}: A list of modules where each module is expected to be a dictionary-like object with a "File" key specifying the file path.

Examples

```julia include_files([Dict("File" => "module1.jl"), Dict("File" => "module2.jl")])