MPI - Functions

Index

PeriLab.IO.find_and_set_core_value_avg

PeriLab.IO.find_and_set_core_value_max

PeriLab.IO.find_and_set_core_value_min

PeriLab.IO.find_and_set_core_value_sum

PeriLab.IO.gather_values

PeriLab.IO.send_single_value_from_vector

PeriLab.IO.send_value

PeriLab.IO.send_vector_from_root_to_core_i

PeriLab.IO.split_vector

PeriLab.IO.synch_controller_bonds_to_responder

PeriLab.IO.synch_controller_bonds_to_responder_flattened

PeriLab.IO.synch_controller_to_responder

PeriLab.IO.synch_responder_to_controller

MPI

PeriLab.IO.send_single_value_from_vector — Function

send_single_value_from_vector(comm::MPI.Comm, controller::Int64, values::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}, type::Type)

Sends a single value from a vector to a controller

Arguments

comm::MPI.Comm: The MPI communicator

controller::Int64: The controller

values::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The values

type::Type: The type

Returns

recv_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The received message

PeriLab.IO.synch_responder_to_controller — Function

synch_responder_to_controller(comm::MPI.Comm, overlapnodes, vector, dof)

Synch the responder to the controller

Arguments

comm::MPI.Comm: The MPI communicator

overlapnodes::Dict: The overlap nodes

vector::Vector: The vector

dof::Int: The degree of freedom

Returns

vector::Vector: The vector

PeriLab.IO.synch_controller_to_responder — Function

synch_controller_to_responder(comm::MPI.Comm, overlapnodes, vector, dof)

Synch the controller to the responder

Arguments

comm::MPI.Comm: The MPI communicator

overlapnodes::Dict: The overlap nodes

vector::Vector: The vector

dof::Int: The degree of freedom

Returns

vector::Vector: The vector

PeriLab.IO.synch_controller_bonds_to_responder — Function

synch_controller_bonds_to_responder(comm::MPI.Comm, overlapnodes, array, dof)

Synch the controller bonds to the responder

Arguments

comm::MPI.Comm: The MPI communicator

overlapnodes::Dict: The overlap nodes

array::Array: The array

dof::Int: The degree of freedom

Returns

array::Array: The array

PeriLab.IO.split_vector — Function

split_vector(input, row_nums, dof)

Split a vector into a vector of matrices

Arguments

input::Vector: The input vector

row_nums::Vector: The row numbers

dof::Int: The degree of freedom

Returns

result::Vector: The result vector

PeriLab.IO.synch_controller_bonds_to_responder_flattened — Function

synch_controller_bonds_to_responder_flattened(comm::MPI.Comm, overlapnodes, array, dof)

Synch the controller bonds to the responder

Arguments

comm::MPI.Comm: The MPI communicator

overlapnodes::Dict: The overlap nodes

array::Array: The array

dof::Int: The degree of freedom

Returns

array::Array: The array

PeriLab.IO.send_vector_from_root_to_core_i — Function

send_vector_from_root_to_core_i(comm::MPI.Comm, send_msg, recv_msg, distribution)

Sends a vector from the root to the core i

Arguments

comm::MPI.Comm: The MPI communicator

send_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The send message

recv_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The receive message

distribution::Vector{Int64}: The distribution

Returns

recv_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The received message

PeriLab.IO.send_value — Function

send_value(comm::MPI.Comm, controller, send_msg)

Sends a value to a controller

Arguments

comm::MPI.Comm: The MPI communicator

controller::Int64: The controller

send_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The send message

Returns

recv_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The received message

PeriLab.IO.find_and_set_core_value_min — Function

find_and_set_core_value_min(comm::MPI.Comm, value::Union{Float64,Int64})

Find and set core value min

Arguments

comm::MPI.Comm: The MPI communicator

value::Union{Float64,Int64}: The value

Returns

recv_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The received message

PeriLab.IO.find_and_set_core_value_max — Function

find_and_set_core_value_max(comm::MPI.Comm, value::Union{Float64,Int64})

Find and set core value max

Arguments

comm::MPI.Comm: The MPI communicator

value::Union{Float64,Int64}: The value

Returns

recv_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The received message

PeriLab.IO.find_and_set_core_value_sum — Function

find_and_set_core_value_sum(comm::MPI.Comm, value::Union{Float64,Int64})

Find and set core value sum

Arguments

comm::MPI.Comm: The MPI communicator

value::Union{Float64,Int64}: The value

Returns

recv_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The received message

PeriLab.IO.find_and_set_core_value_avg — Function

find_and_set_core_value_avg(comm::MPI.Comm, value::Union{Float64,Int64})

Find and set core value avg

Arguments

comm::MPI.Comm: The MPI communicator

value::Union{Float64,Int64}: The value

Returns

recv_msg::Union{Int64,Vector{Float64},Vector{Int64},Vector{Bool}}: The received message

PeriLab.IO.gather_values — Function

gather_values(comm::MPI.Comm, value::Any)

Gather values

Arguments

comm::MPI.Comm: The MPI communicator

value::Any: The value

Returns

recv_msg::Any: The received message