Getting Started

Installation

The PeriLab package is available through the Julia package system and can be installed using the following commands:

using Pkg
Pkg.add("PeriLab")

Throughout the rest of this tutorial, we will assume that you have installed the PeriLab package and have already typed using PeriLab to bring all of the relevant variables into your current namespace.

Using PeriLab

The simplest way to run the PeriLab simulation core is to use a provided example import the module and go.

using PeriLab

PeriLab.get_examples()
PeriLab.main("examples/DCB/DCBmodel.yaml")

The output should look like this:

The main functionalities for the yaml input deck is given in

"examples/functionalities.yaml"

Using PeriLab with multiple processors (MPI)

In order to run PeriLab for large scale problems MPI needs to be installed:

$ julia
julia> using MPI
julia> MPI.install_mpiexecjl()

Run PeriLab with two processors:

$ mpiexecjl -n 2 julia -e "using PeriLab; PeriLab.main()" Dogbone.yaml -v

Training

The training input is given under the examples folder. The documentation and a video will follow.

Index

Functions

PeriLab.main — Function

main()

Entry point for the PeriLab application.

This function serves as the entry point for the PeriLab application. It calls the core main function with the provided arguments.

main(filename::String, output_dir::String="", dry_run::Bool=false, verbose::Bool=false, debug::Bool=false, silent::Bool=false, reload::Bool=false)

Entry point for the PeriLab application.

This function serves as the entry point for the PeriLab application. It calls the core main function with the provided arguments.

Arguments

filename::String: The filename of the input file.

output_dir::String: The output directory.

dry_run::Bool=false: Whether to run in dry-run mode.

verbose::Bool=false: Whether to run in verbose mode.

debug::Bool=false: Whether to run in debug mode.

silent::Bool=false: Whether to run in silent mode.

reload::Bool=false: Whether to reload the input file.

PeriLab.get_examples — Function

get_examples()

Copy the examples folder to the current directory.